WE HAZ FONTS

+++ YA3DE/YA3DE/Game.cpp
@@ -130,7 +130,7 @@ void Game::Run()
 
 		_Window.EndScene();
 
-		elapsed = std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::high_resolution_clock::now() - start).count();
+		elapsed = (double)std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::high_resolution_clock::now() - start).count();
 	}
 	
 	LOG("Unloading");
